requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.1.0
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.7
django-autoslug==1.9.8
django-extensions==3.2.1
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
export==0.2.0
factory-boy==3.2.1
Faker==17.0.0
flake8==6.1.0
flake8-import-order==0.18.2
gunicorn==20.1.0
html2text==2020.1.16
ipdb==0.13.11
ipython==8.11.0
jedi==0.18.2
kombu==5.2.4
matplotlib-inline==0.1.6
mccabe==0.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
